Reworked PowerManagement. - Multiple Power Provider in a Pipe now provides Multiamps - Reworked CodeStructure to remove duplicate checks and code - Implemented GTProxy as some GTBlocks do not directly implement the IC2 IEnergySink Interface

+++ src/main/java/logisticspipes/pipes/basic/CoreRoutedPipe.java
@@ -1565,11 +1565,11 @@ public abstract class CoreRoutedPipe extends CoreUnroutedPipe
     }
 
     public void handleRFPowerArival(float toSend) {
-        powerHandler.addRFPower(toSend);
+        powerHandler.addRFPower((long) toSend);
     }
 
     public void handleIC2PowerArival(float toSend) {
-        powerHandler.addIC2Power(toSend);
+        powerHandler.addIC2Power((long) toSend);
     }
 
     /*** IInventoryProvider ***/
