+++ src/main/java/logisticspipes/pipes/upgrades/power/IC2EVPowerSupplierUpgrade.java
@@ -3,7 +3,7 @@ package logisticspipes.pipes.upgrades.power;
 public class IC2EVPowerSupplierUpgrade extends IC2PowerSupplierUpgrade {
 
     @Override
-    public int getPowerLevel() {
+    public long getPowerLevel() {
         return 2048;
     }
 }
