+++ src/main/java/logisticspipes/pipes/upgrades/power/IC2LVPowerSupplierUpgrade.java
@@ -3,7 +3,7 @@ package logisticspipes.pipes.upgrades.power;
 public class IC2LVPowerSupplierUpgrade extends IC2PowerSupplierUpgrade {
 
     @Override
-    public int getPowerLevel() {
+    public long getPowerLevel() {
         return 32;
     }
 }
